Added slider which can be used to seek on the track.
As the project is using Qt 4 it was not possible to just use
a out of the box slider. The example was taken from here and
modified: https://doc.qt.io/archives/qt-4.8/qt-demos-declarative-flickr-qml-flickr-common-slider-qml.html

Qt5 and QtQuick 2 already have a slider built in. It would be
a good idea to update to get the new features. I will need to
consult to findout if this would break any builds before updating.